# doc/openapi.yaml
openapi: 3.0.3
info:
  title: SideQuestCommunity API
  version: "1.0.0"
  description: OpenAPI spec generated from backend controllers.
servers:
  - url: /
...
paths:
...
  /api/core/posts:
    get:
      tags: [core]
      summary: List posts
      parameters:
        - name: current
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sectionId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultPagePostVO"
    post:
      tags: [core]
      summary: Create post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultLong"
  /api/core/posts/following:
    get:
      tags: [core]
      summary: List posts from following
      security:
        - bearerAuth: []
      parameters:
        - name: current
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultPagePostVO"
...
  /api/core/posts/{id}:
    get:
      tags: [core]
      summary: Get post by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultPostVO"
  /api/core/admin/posts/sync:
    post:
      tags: [admin, core]
      summary: Sync posts to search
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultString"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResultString:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data: { type: string, nullable: true }
...
    ResultLong:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data: { type: integer, format: int64, nullable: true }
...
    ResultPostVO:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          $ref: "#/components/schemas/PostVO"
    ResultPagePostVO:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          $ref: "#/components/schemas/PagePostVO"
...
    PagePostVO:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/PostVO"
        total: { type: integer, format: int64 }
        size: { type: integer }
        current: { type: integer }
        pages: { type: integer }
...
    CreatePostDTO:
      type: object
      required: [title, content, sectionId]
      properties:
        title: { type: string }
        content: { type: string }
        sectionId: { type: integer, format: int64 }
        tags:
          type: array
          items: { type: string }
        imageUrls:
          type: array
          items: { type: string }
        videoUrl: { type: string, nullable: true }
        videoCoverUrl: { type: string, nullable: true }
        videoDuration: { type: integer, nullable: true }
...
    PostVO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        authorId: { type: integer, format: int64 }
        authorName: { type: string }
        authorAvatar: { type: string }
        title: { type: string }
        content: { type: string }
        sectionId: { type: integer, format: int64 }
        status: { type: integer }
        likeCount: { type: integer }
        commentCount: { type: integer }
        favoriteCount: { type: integer }
        viewCount: { type: integer }
        createTime: { type: string, format: date-time }
        updateTime: { type: string, format: date-time }
        imageUrls:
          type: array
          items: { type: string }
        videoUrl: { type: string, nullable: true }
        videoCoverUrl: { type: string, nullable: true }
        videoDuration: { type: integer, nullable: true }
        tags:
          type: array
          items: { type: string }
        hasLiked: { type: boolean }
        hasFavorited: { type: boolean }
        isFollowing: { type: boolean }
...